+++ lib/database/entities/history/stream_state_entity.dart
@@ -3,7 +3,7 @@ import 'package:floor/floor.dart';
 @Entity(tableName: StreamStateEntity.tableName)
 class StreamStateEntity {
   @primaryKey
-  int streamId;
+  String streamId;
   int progressTime;
 
   StreamStateEntity(this.streamId, this.progressTime);
